jarvis/backend/requirements.txt
flask
flask-socketio
python-dotenv
requests